requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
passlib[bcrypt]
python-jose
email-validator
apscheduler